# openapi/LupiraCalApi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Lupira Cal API",
    "description": "Calendar backend for Lupira (contacts live in LupiraContactApi; the DAV protocol surface in LupiraDavApi). Authenticate with a Bearer token issued by the OIDC provider (Authentik).",
    "version": "v1"
  },
  "paths": {
...
    "/calendars/{calendarId}/proposed": {
      "get": {
        "tags": [
          "Curation"
        ],
        "summary": "List items proposed into a calendar (awaiting accept/reject).",
        "operationId": "ListProposedItems",
        "parameters": [
          {
            "name": "calendarId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CalendarItemDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/items/{itemId}/calendars/{calendarId}/accept": {
      "post": {
        "tags": [
          "Curation"
        ],
        "summary": "Accept a proposed item into a calendar. Offline clients pass ?occurredAt= + Idempotency-Key for replay-safe, last-writer-wins filing.",
        "operationId": "AcceptItemIntoCalendar",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "calendarId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "occurredAt",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarItemDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/items/{itemId}/calendars/{calendarId}": {
      "post": {
        "tags": [
          "Curation"
        ],
        "summary": "File an existing item into a calendar (status=proposed|accepted, default proposed). Offline clients pass ?occurredAt= + Idempotency-Key.",
        "operationId": "FileItemToCalendar",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "calendarId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "occurredAt",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarItemDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Curation"
        ],
        "summary": "Remove an item from a calendar (reject / unfile). Offline clients pass ?occurredAt= + Idempotency-Key.",
        "operationId": "RemoveItemFromCalendar",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "calendarId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "occurredAt",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarItemDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/sync/changes": {
      "get": {
        "tags": [
          "Sync"
        ],
        "summary": "Delta feed for offline mirrors: everything the caller can read that changed past the cursor, plus tombstone ids for items deleted or no longer visible. Omit since for a full sync (tombstones suppressed — replace the mirror wholesale); loop while hasMore, persisting cursor between calls.",
        "operationId": "GetChanges",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncChangesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/sync/containers": {
      "get": {
        "tags": [
          "Sync"
        ],
        "summary": "Snapshot of the caller's calendars for mirror reconciliation. Containers are plain documents with no event history (no cursor) — fetch once per sync cycle and diff locally.",
        "operationId": "GetSyncContainers",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncContainersResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "Access": {
        "enum": [
          "Owner",
          "ReadWrite",
          "Read"
        ]
      },
      "ActionKind": {
        "enum": [
          "SendCheckIn",
          "Notify",
          "CreateLinkedTask",
          "ExpireTarget",
          "RescheduleSelf",
          "RunJob",
          "Rescore"
        ]
      },
...
      "BookingDetail": {
        "required": [
          "providerContactId",
          "confirmationNumber",
          "reference",
          "url",
          "amount",
          "currency",
          "partySize"
        ],
        "type": "object",
        "properties": {
          "providerContactId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "confirmationNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "reference": {
            "type": [
              "null",
              "string"
            ]
          },
          "url": {
            "type": [
              "null",
              "string"
            ]
          },
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "currency": {
            "type": [
              "null",
              "string"
            ]
          },
          "partySize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "CalendarClass": {
        "enum": [
          "Agenda",
          "System",
          null
        ]
      },
      "CalendarEntryStatus": {
        "enum": [
          "Proposed",
          "Accepted",
          "Removed"
        ]
      },
      "CalendarItemDto": {
        "required": [
          "id",
          "externalId",
          "isAllDay",
          "attendees",
          "calendars",
          "etag",
          "createdAt",
          "updatedAt",
          "version"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "externalId": {
            "type": "string"
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ItemStatus"
              }
            ]
          },
          "isAllDay": {
            "type": "boolean"
          },
          "startsAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "endsAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "startDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "endDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "startPrecision": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DatePrecision"
              }
            ]
          },
          "endPrecision": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DatePrecision"
              }
            ]
          },
          "recurrenceRule": {
            "type": [
              "null",
              "string"
            ]
          },
          "category": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ItemCategory"
              }
            ]
          },
          "details": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ItemDetails"
              }
            ]
          },
          "placeId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "locationLabel": {
            "type": [
              "null",
              "string"
            ]
          },
          "parentItemId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "tags": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "metadata": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/JsonNode"
              }
            ]
          },
          "prompt": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ItemPrompt"
              }
            ]
          },
          "action": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ItemAction"
              }
            ]
          },
          "completeness": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CompletenessScore"
              }
            ]
          },
          "attendees": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemAttendee"
            }
          },
          "calendars": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CalendarMembershipDto"
            }
          },
          "etag": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "CalendarKind": {
        "enum": [
          "Personal",
          "Group",
          "Birthdays",
          "Availability",
          "Inbox",
          "LlmPrompts",
          "UserCheckIn",
          "DevOps",
          "FoodPlan",
          "Generic",
          null
        ]
      },
      "CalendarMembershipDto": {
        "required": [
          "calendarId",
          "status"
        ],
        "type": "object",
        "properties": {
          "calendarId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/CalendarEntryStatus"
          }
        }
      },
      "CompletenessGap": {
        "required": [
          "field",
          "weight",
          "severity"
        ],
        "type": "object",
        "properties": {
          "field": {
            "type": "string"
          },
          "weight": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "severity": {
            "$ref": "#/components/schemas/GapSeverity"
          }
        }
      },
      "CompletenessScore": {
        "required": [
          "score",
          "rubricVersion",
          "gaps"
        ],
        "type": "object",
        "properties": {
          "score": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "rubricVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "gaps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompletenessGap"
            }
          }
        }
      },
      "ContainerDto": {
        "required": [
          "id",
          "type",
          "slug",
          "access"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "displayName": {
            "type": [
              "null",
              "string"
            ]
          },
          "color": {
            "type": [
              "null",
              "string"
            ]
          },
          "defaultTimezone": {
            "type": [
              "null",
              "string"
            ]
          },
          "class": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CalendarClass"
              }
            ]
          },
          "kind": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CalendarKind"
              }
            ]
          },
          "access": {
            "$ref": "#/components/schemas/Access"
          }
        }
      },
...
      "DatePrecision": {
        "enum": [
          "Exact",
          "Day",
          "Month",
          "Year",
          "Approximate",
          null
        ]
      },
      "FallbackMode": {
        "enum": [
          "Retry",
          "Ask",
          "Drop"
        ]
      },
...
      "ItemAction": {
        "required": [
          "kind",
          "target",
          "paramsJson",
          "fire",
          "enabled"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/ActionKind"
          },
          "target": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Ref"
              }
            ]
          },
          "paramsJson": {
            "type": "string"
          },
          "fire": {
            "$ref": "#/components/schemas/PromptFire"
          },
          "enabled": {
            "type": "boolean"
          }
        }
      },
      "ItemAttendee": {
        "type": "object",
        "properties": {
          "participationId": {
            "type": "string",
            "format": "uuid"
          },
          "contactId": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "$ref": "#/components/schemas/ParticipationRole"
          },
          "status": {
            "$ref": "#/components/schemas/ParticipationStatus"
          },
          "invitedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "respondedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "attendedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "leftAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "ItemCategory": {
        "enum": [
          "General",
          "Meeting",
          "Appointment",
          "Meal",
          "Occasion",
          "Outing",
          "Trip",
          "Stay",
          "Activity",
          "Focus",
          "Chore",
          null
        ]
      },
      "ItemDetails": {
        "type": "object",
        "properties": {
          "booking": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BookingDetail"
              }
            ]
          },
          "travel": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TravelLeg"
              }
            ]
          },
          "presence": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PresenceDetail"
              }
            ]
          }
        }
      },
...
      "ItemPrompt": {
        "required": [
          "intent",
          "target",
          "instruction",
          "output",
          "tools",
          "tier",
          "onMiss",
          "fire",
          "enabled"
        ],
        "type": "object",
        "properties": {
          "intent": {
            "$ref": "#/components/schemas/PromptIntent"
          },
          "target": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Ref"
              }
            ]
          },
          "instruction": {
            "type": "string"
          },
          "output": {
            "$ref": "#/components/schemas/OutputKind"
          },
          "tools": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "tier": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ModelTier"
              }
            ]
          },
          "onMiss": {
            "$ref": "#/components/schemas/FallbackMode"
          },
          "fire": {
            "$ref": "#/components/schemas/PromptFire"
          },
          "enabled": {
            "type": "boolean"
          }
        }
      },
      "ItemStatus": {
        "enum": [
          "Tentative",
          "Confirmed",
          "Cancelled",
          null
        ]
      },
      "JsonNode": { },
...
      "ModelTier": {
        "enum": [
          "Small",
          "Medium",
          "Large",
          null
        ]
      },
...
      "OutputKind": {
        "enum": [
          "RecordEdit",
          "Event",
          "Task",
          "Message",
          "Summary",
          "Question",
          "Relation",
          "None"
        ]
      },
...
      "ParticipationRole": {
        "enum": [
          "Chair",
          "RequiredParticipant",
          "OptionalParticipant",
          "NonParticipant"
        ]
      },
      "ParticipationStatus": {
        "enum": [
          "NeedsAction",
          "Accepted",
          "Declined",
          "Tentative",
          "Delegated"
        ]
      },
...
      "PresenceDetail": {
        "required": [
          "status"
        ],
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/AvailabilityStatus"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "PromptFire": {
        "required": [
          "kind",
          "offsetMinutes",
          "allDayAt"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/PromptFireKind"
          },
          "offsetMinutes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "allDayAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "time"
          }
        }
      },
...
      "PromptIntent": {
        "enum": [
          "EnrichRecord",
          "Research",
          "CreateFollowUp",
          "Monitor",
          "Summarise",
          "AskUser"
        ]
      },
      "Ref": {
        "required": [
          "kind",
          "id",
          "url"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/RefKind"
          },
          "id": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "url": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SectionGuardsDto": {
        "required": [
          "core",
          "metadata",
          "payload",
          "filing"
        ],
        "type": "object",
        "properties": {
          "core": {
            "$ref": "#/components/schemas/SectionGuardDto"
          },
          "metadata": {
            "$ref": "#/components/schemas/SectionGuardDto"
          },
          "payload": {
            "$ref": "#/components/schemas/SectionGuardDto"
          },
          "filing": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/SectionGuardDto"
            }
          }
        }
      },
...
      "SyncChangeDto": {
        "required": [
          "item",
          "guards"
        ],
        "type": "object",
        "properties": {
          "item": {
            "$ref": "#/components/schemas/CalendarItemDto"
          },
          "guards": {
            "$ref": "#/components/schemas/SectionGuardsDto"
          }
        }
      },
      "SyncChangesResponse": {
        "required": [
          "cursor",
          "hasMore",
          "changed",
          "deleted"
        ],
        "type": "object",
        "properties": {
          "cursor": {
            "type": "string"
          },
          "hasMore": {
            "type": "boolean"
          },
          "changed": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SyncChangeDto"
            }
          },
          "deleted": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
      "SyncContainersResponse": {
        "required": [
          "calendars"
        ],
        "type": "object",
        "properties": {
          "calendars": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContainerDto"
            }
          }
        }
      },
      "TransportMode": {
        "enum": [
          "Flight",
          "Train",
          "Metro",
          "Tram",
          "Bus",
          "Coach",
          "Car",
          "Ferry",
          "Bike",
          "Walk",
          "Other"
        ]
      },
      "TravelLeg": {
        "required": [
          "mode",
          "toPlaceId",
          "fromPlaceId",
          "departAt",
          "arriveAt",
          "carrier",
          "serviceNumber",
          "departurePoint",
          "arrivalPoint",
          "seat",
          "driverContactId"
        ],
        "type": "object",
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/TransportMode"
          },
          "toPlaceId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "fromPlaceId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "departAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "arriveAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "carrier": {
            "type": [
              "null",
              "string"
            ]
          },
          "serviceNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "departurePoint": {
            "type": [
              "null",
              "string"
            ]
          },
          "arrivalPoint": {
            "type": [
              "null",
              "string"
            ]
          },
          "seat": {
            "type": [
              "null",
              "string"
            ]
          },
          "driverContactId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "toLabel": {
            "type": [
              "null",
              "string"
            ]
          },
          "fromLabel": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "OIDC bearer token. Send as `Authorization: Bearer <token>`.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}